requirements.txt
langchain
langgraph
faiss-cpu
openai
python-dotenv